Refactor modules, create separate GEMM and GEMV
