--- awq/modules/__init__.py
@@ -1,3 +1,0 @@
-from .fused_norm import *
-from .fused_attn import *
-from .fused_mlp import *
